예제 드라이버 변경.

+++ src/main/java/org/openflamingo/mapreduce/util/correlation/PearsonCorrelation.java
@@ -15,23 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.openflamingo.mapreduce.util.correlation;
 
 /**
@@ -49,6 +32,7 @@ public class PearsonCorrelation {
         double sum_coproduct = 0;
         double mean_x = scores1[0];
         double mean_y = scores2[0];
+
         for (int i = 2; i < scores1.length + 1; i += 1) {
             double sweep = Double.valueOf(i - 1) / i;
             double delta_x = scores1[i - 1] - mean_x;
@@ -59,8 +43,9 @@ public class PearsonCorrelation {
             mean_x += delta_x / i;
             mean_y += delta_y / i;
         }
-        double pop_sd_x = (double) Math.sqrt(sum_sq_x / scores1.length);
-        double pop_sd_y = (double) Math.sqrt(sum_sq_y / scores1.length);
+
+        double pop_sd_x = Math.sqrt(sum_sq_x / scores1.length);
+        double pop_sd_y = Math.sqrt(sum_sq_y / scores1.length);
         double cov_x_y = sum_coproduct / scores1.length;
         result = cov_x_y / (pop_sd_x * pop_sd_y);
         return result;
